pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"
description = "basic todo crud app"
authors = ["Omar622 <omaramin622@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"
pylint = "^3.0.2"
pylint-django = "^2.5.5"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
